requirements.txt
Flask
Werkzeug
cryptography
pyOpenSSL
TinyDB
